requirements.txt
imageio==2.5.0
scikit-image==0.15.0
numpy==1.17.0
cairocffi==1.1.0
ndjson==0.3.1
Pillow==6.1.0